pyproject.toml
[tool.poetry]
name = "repl-nix-projectexplorer"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [
    { include = "app" }
]

[tool.poetry.dependencies]
python = "^3.11"
flask = "^3.0.3"
groq = "^0.23.1"
google-generativeai = "^0.8.1"
anthropic = "^0.34.2"
openai = "^1.76.2"
cerebras-cloud-sdk = "^1.3.0"
python-dotenv = "^1.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
